refactor(storage): move proximablocks spatial_* cohort to engine-core (#998)
Hoists the spatial-clustering cohort (spatial_clustering + spatial_encoding +
spatial_pruning + spatial_traits, ~2,329 LOC) out of the root crate into
proximadb-engine-core. Third increment of the engines::core keystone extraction
(after simd_decode #990 + header_metadata/per_column_alignment #997).

The cohort is self-contained (super:: refs only to each other); the single
external ref — crate::storage::engines::helix::hilbert_curve::HilbertCurve —
swaps to proximadb_storage_common::hilbert_curve::HilbertCurve (the helix file
is just a re-export shim for the storage_common type). Root keeps per-file
re-export shims; all ...::formats::proximablocks::spatial_*::* paths resolve
unchanged.

Crate gains dep: proximadb-clustering-kernel (spatial_clustering k-means).
distance-kernel + storage-common already present.
Verified: crate lib + tests (78/78), root check --lib + --tests (authoritative
full-clean), clippy --lib --bins -D warnings, fmt, layering, workspace
boundaries, storage up-edges (<=180), work-commit-check.
