+++ crates/storage/proximadb-engine-core/src/proximablocks/spatial_encoding.rs
@@ -1,0 +1,2 @@
+//! Compatibility shim — implementation now lives in `proximadb-storage-common`.
+pub use proximadb_storage_common::spatial_encoding::*;
