+++ src/storage/engines/core/formats/proximablocks/spatial_encoding.rs
@@ -1,2 +1,5 @@
-//! Compatibility shim — implementation now lives in `proximadb-storage-common`.
-pub use proximadb_storage_common::spatial_encoding::*;
+// Copyright (C) 2025 ProximaDB
+// SPDX-License-Identifier: Apache-2.0
+//! Re-export shim — `spatial_encoding` hoisted to `proximadb-engine-core`.
+//! All `...::formats::proximablocks::spatial_encoding::*` paths resolve unchanged.
+pub use proximadb_engine_core::proximablocks::spatial_encoding::*;
